+++ CHANGELOG.rst
@@ -2,6 +2,19 @@
 Changelog
 =========
 
+Version 0.5.0 (2026-02-12)
+===========
+
+- Try, Except blocks introduced to prevent the calculation loop to stop (Issue 42).
+- An structlog logger with two handlers (file and console) was introduced and print statements were removed.
+- The flag `--verbose` was introduced to control the verbosity of the logger.
+- The logic of the EPDs calculation in `src/epd/pipeline.py` was changed. The `gen_epds` generator is now created just once making the mode more efficient.
+
+Version 0.4.0 (2025-12-05)
+===========
+
+- MISSING
+
 Version 0.3.0 (2025-11-12)
 ===========